requirements.txt
istari-digital-client
python-dotenv
